+++ CHANGELOG.rst
@@ -8,14 +8,18 @@ Notable project changes since release 1.3.1 (2017-02-22).
 Unreleased
 ==========
 
+
 Added
 -----
 
+- DFTB3 coupled-perturbed atom derivatives
+
+- Derivatives with respect to external charges, controlled both via
+  input and API
 
 
 Changed
 -------
-
 
 
 Fixed